+++ server/db/connections.tsx
@@ -1,9 +1,9 @@
 import { MongoClient } from "mongodb";
 import { Sequelize } from "sequelize-typescript";
 import { mongoConfig, psqlConfig } from "./configs";
-import SiteTheme from "./models/schemes/site-theme";
-import User from "./models/schemes/user";
-import UserTheme from "./models/schemes/user-theme";
+import SiteTheme from "./schemes/site-theme";
+import User from "./schemes/user";
+import UserTheme from "./schemes/user-theme";
 
 const psqlConnection = new Sequelize(psqlConfig);
 psqlConnection.addModels([UserTheme, SiteTheme, User]);
